--- 1.rb
@@ -1,4 +1,0 @@
-ootsukadesu
-yoshidadesu.
-
-fjsdfkajslfjkladsjfkljaskljfkjdsakjfklasjdklfjksdajfjkdljk
