--- 2.rb
@@ -1,2 +1,0 @@
-こんにちは
-# backup
