+++ README.md
@@ -1,0 +1,16 @@
+# 自販機プログラム
+
+> ## ファイルの構成
+>> * vending_machine.rb  
+>>> class:VendingMachineのファイル
+>> * item_setting.rb  
+>>> class:ItemSetのファイル
+>> * modules.rb
+>>> module:MachineHelperのファイル
+> ## クラス・モジュールの内容
+>> * VendingMachine  
+>>> 自販機の挙動に関するコードの集まり
+>> * ItemSet  
+>>> 在庫・売上金額などの管理に関するコードの集まり
+>> * MachineHelper  
+>>> 交通の処理をまとめたモジュール
